tests/pty.sh: Fix fail for long $TMPDIR or narrow terminal window
For a sufficiently narrow Terminal.app window, the test on line 999
fails consistently. The expected tab-completed pathname is cut off.

In spite of pty being a pseudoterminal, ksh gets the window size
from the real terminal. I don't know how to fix that.

By default, the pty tests are run with TERM=dumb, on which
multiline editing is not possible. The temporary file path may not
fit on a single line.

The fix: run the test with multiline enabled.

Resolves: https://github.com/ksh93/ksh/issues/653