emacs: Suppress hist output to command line with ^X^E (#659)
The function ed_fulledit operates by writing the appropriate hist
command to the input buffer. In emacs mode, this results in the
full hist command being printed to the command line and remaining
there. This behavior does not occur in other shells, and can safely
be considered a bug. The output can be suppressed, however, by not
using draw with the option FINAL after ^X^E:

src/cmd/ksh93/edit/emacs.c:
- Add a variable ehist to _emacs_ and initialize to 0; it is set to
  1 by the ^X^E command, suppressing the use of draw(ep,FINAL) in
  ed_emacsread. This prevents the altered input buffer from being
  displayed onscreen.

Resolves: https://github.com/ksh93/ksh/issues/563