Avoid race condition in comsub redirection (re: 274def65, e373e8c1)
Turns out that running sh_subfork() when redirecting the output
of external commands in command substitutions triggers a race
condition on some systems. So we need to limit the workaround to
built-ins, without reintroducing previous problems.

src/cmd/ksh93/sh/io.c: sh_redirect():
- To limit the workaround, use the sh.redir0 flag set in xec.c --
  it's for an optimisation of the 'read' command (introduced in ksh
  93u+ 2011-12-24), but we can use the fact that xec.c sets it to 1
  for any built-in command right before calling sh_redirect(). This
  covers permanent redirections with 'exec' and 'redirect' as well.
- Check all the redirections for need to fork, in a separate loop.

Thanks to @phidebian for the report.

Resolves: https://github.com/ksh93/ksh/issues/688
Resolves: https://github.com/ksh93/ksh/pull/690