kill: re-allow sending extra SIGCONT as documented (re: a20b7700)
I was a little hasty in removing that behaviour entirely. According
to the documentation, sending an extra SIGCONT is expected if, and
only if, the signal is SIGHUP or SIGTERM *and* the process in
question is currently stopped:

sh.1: "If the signal being sent is TERM (terminate) or HUP
(hangup), then the job or process will be sent a CONT (continue)
signal if it is stopped." The Bolsky & Korn book has similar.

This commit reimplements this behaviour in a way that acts as
documented. However, it still violates POSIX, so the posix shell
option disables the sending of an extra SIGCONT altogether.

src/cmd/ksh93/sh/jobs.c:
- Add a helper function also_send_sigcont() that checks if the
  POSIX mode is off, the signal is SIGHUP or SIGTERM, and the job
  associated with the PID is flagged as P_STOPPED.
- Use that function in job_kill() to check whether to send an extra
  SIGCONT (in four different places).
- job_unstop(): Be better behaved. Add a flag indicating whether to
  actually send SIGCONT or only mark the job as unstopped in the
  internal jobs table. This allows job_kill() to avoid sending
  SIGCONT twice when resuming a process group.
- Another old bug fixed in job_kill(): when processing a negative
  PID (which sends a signal to that PID's process group), the job
  entry was not obtained, as job_bypid() was given the negative PID
  from pw->p_pid (== pid), so job_unstop() got a NULL pointer.
  Fixed by pid = -pid and making the code generally more legible.