package: check for changes in $CC, etc. earlier (re: 48e6dd98)
The check if $CC was changed did not work because it came after
code that stored $CC in the generated bin/cc script and then set
the CC variable to 'cc'. The fix is to check before that. The
corresponding code block is moved accordingly. Some tweaks for
robustness are also applied.

Unfortunately, after this commit, if you are using a build
directory where you had set CC to a custom compiler path, you'll
now get a spurious error that it has changed from 'cc' because the
bug caused it to remmeber that value. To fix this, remove the file
arch/*/lib/package/gen/CC.