[v1.1] remove explicit_bzero(3) prototype probe (re: 00b296cd)
On systems where explicit_bzero(3) exists (i.e., on Linux), the
extern declaration will be in strings.h, so the ntp probe will not
generate a prototype declaration. On systems where it doesn't
exist, we don't ship our own copy and don't use it (it's #define'd
in arc4random.c to a call to a volatile pointer to memset(3)), so
there is no point in generating the prototype declaration.