[v1.1] Save/restore SRANDOM in virtual subshells (re: 00b296cd) (#710)
This commit fixes a virtual subshell leak that caused the set upper
bound for SRANDOM to leak out into the parent shell. This is fixed
by saving and restoring the parent shell's upper bound for SRANDOM
(the upper bound variable has been moved to the sh struct for this
purpose).