Fix bad default 'return' status in traps (re: aea99158, 092b90da)
As of aea99158, the src/cmd/ksh93/fun/autocd autoloadable function
is broken. After activating autocd and changing directory with just
the directory name (causing exit status 2), any command with more
than one argument (i.e.: containing a blank) would simply not be
executed (and the exit status would stay 2).

$ autocd
autocd activated -- type a directory path to cd, ~- to cd back
type 'autocd -d' to deactivate
$ /
cd -- /
$ ls bin/
$ echo what
$ true
$ echo what
what

After a command with a single argument, normal operation is
restored until the next time autocd is triggered.

The cause is incorrect behaviour of the 'return' command on line 31
in fun/autocd. That command should return from the _do_autocd()
function with the default status 0. Instead, it returns with
whatever status was in $? just before the trap was invoked, which
is wrong. In this case, that is 2, since we just did an autoce.
This in turn causes the DEBUG trap to stop execution of the command
entered (a documented feature which I used to make autocd stop
directories from being executed as commands in the first placae).

Unfortunately, this bug made it into the 1.0.8 release. To make
autocd bug-compatible with ksh 93u+m/1.0.8, change that 'return'
to an explicit 'return 0'.

So, short version: the default status of 'return' and 'exit', with
no argument, is broken within traps. They need to use the status of
the last-executed command, even within traps. It's only the trap
itself, as a whole, that needs to exit with the status saved from
before trap execution.

Analysis:

This is an old AT&T bug that was fixed in 092b90da and reintroduced
in aea99158. In 092b90da, I was right in the first place: that
whole sh.oldexit mess is broken and it was correct to remove it. A
different fix is needed for the issue it attempted to solve.

sh_trap() already contains code to save the exit status in a local
variable. All we should need beyond that is to know whether 'exit'
was invoked with an explicit exit status. In that case, the trap
needs to exit with that status. In all other cases, the trap as a
whole needs to exit with the saved pre-trap status, without
affecting 'exit' or 'return' *within* the trap. That should be
enough to keep the regression tests added in aea99158 passing.

This commit reverts aea99158 and instead applies the following fix:

src/cmd/ksh93/include/shell.h:
- Add 'sh.intrap_exit_n' flag.

src/cmd/ksh93/bltins/cflow.c: b_return() (which also does 'exit'):
- When in a trap, and an explicit exit status argument was given to
  'exit', then set the 'sh.intrap_exit_n' flag.

src/cmd/ksh93/sh/fault.c: sh_trap():
- Remove broken logic that tries to decide whether to restore the
  pre-trap exit status (savxit) based on jmpval. This may show
  whether exit was used, but not whether it was given an explicit
  exit status argument.
- Replace that with a check for sh.intrap_exit_n. Only restore the
  saved status if that flag was not set.