[v1.1]: rm init of sh.tilde discipline (re: 936a1939)
The two removed lines simply make it possible to define a function
called '.sh.tilde'. They otherwise don't do anything. This is not
necessary for the ksh 93u+m sh.tilde.{get,set} disciplines.

Just in case, this is removed for 1.1+ only and kept on 1.0 for
backward compatibility (as unlikely as it is that this is needed).