mamake: bind: also bind to local .a files (re: 1400c27a, c3272bc2)
The first referenced commit exposed a mamake bug introduced in the
second: when executing 'bind', a dependency was only declared on a
file that has an absolute path -- effectively, one that is already
preinstalled. But *.a files that exist in the current working
directory also need to count; when we link a binary against a
library we just built, such as linking ksh against libshell, that's
what ${mam_libshell} contains.

To fix it, instead of only declaring dependencies on /* paths,
declare dependencies on everything except strings that start with
'-l' (which are compiler flags, not real files).