sfio: rename all SF_* to SFIO_* (re: bb0c31e1, 4dcf5c50, 2cf1ffd0)
In include/sfio.h, sfio/sfhdr.h and include/ast_windows.h, the code
contains several workarounds for conflicts with SF_* macros defined
in sys/socket.h on various systems. More problems may occur in
future as there is no guarantee that some systems' socket.h won't
add more macros that may conflict with ours.

Commit 4dcf5c50 already partially gave up on backward compatibility
by renaming SF_FLAGS to SFIO_FLAGS (and making SF_FLAGS the
workaround added in 2cf1ffd0 redundant).

There is only one way to make this future-proof: forget about
backward compatibility completely and rename all the SF_* bitmask
and event macros to SFIO_*. This also allows us to get rid of those
workarounds.

This commit also bumps SFIO_VERSION (as well as AST_VERSION) to
20240303L (note that SFIO_VERSION was never called SF_VERSION). So
if anyone ever needs to write code that is compatible with both
SFIO_* and SF_*, it is possible with a preprocessor check for
SFIO_VERSION < 20240303L.

*** Build system fix ***

The above changes exposed a bug in libast/Mamfile: when Cdt was
building on my system after applying the changes, the compiler
warned that sfio.h was redefining the SF_APPEND from sys/stat.h,
though there was no trace of SF_APPEND left in sfio.h. The cause
was an outdated copy of sfio.h left in the object directory from
the previous build; Cdt gets (re)built before Sfio. This was
included because of an erroneous -I${PACKAGEROOT}/include/ast in
the compiler invocation. Removing that fixes the problem.