mamake: tweak for future-proofing
The search() function now always returns a pointer to the
variable/dictionary item/node and not to its value; setval() and
getval() functions are added to deal with values easily. This
should make it a bit easier to add some features down the line.

Also a minor de-obfuscation: get rid of the ROTATE() macro
in search() and just do those assignments directly.