[v1.1] break/continue optimization: fix # args check (re: 2e1e47d6)
Using argchn.len to check for the number of command arguments was
completely bogus; in fact, that union member of struct argnod was
unused. It worked on 64-bit systems but that is purely accidental.

Studying sh_argbuild() in args.c shows us how to do this correctly.
When a command has no expansions (so the COMSCAN bit is not set),
the argument takes the form of 'struct dolnod' values even if
though is pointed to by a pointer of tupe 'struct argnod'. The
dolnum member contains the number of arguments, starting at 1 for
the command name.

The reason argchn.len in struct argnod accidentally worked on
64-bit systems is that its offset in struct argnod (where it
follows two 32-bit ints) happens to be equal to the offset of
dolnum in struct dolnod (where it follows one 64-bit pointer).
Of course that is not the case on systems where pointers are
not 64 bits long.

The argchn.len union member is unused as of 66b6ce4e and is now
removed to avoid future confusion.