struct comnod: turn .comarg into a union
The comarg member of struct comnod is declared as a pointer of type
struct argnod, but (as the code in sh_argbuild() and parse.c shows)
this only applies before arguments are expanded (when tretyp has
the COMSCAN bit set). After expansion or for unexpandable arguments
(no COMSCAN bit set), that pointer actually points to a list of
type struct dolnod, a data structure also used for positional
parameters. So the code contains a fair bit of typecasting on comarg
to cope with this.

Unions are in C for this sort of situation. So this commit turns
comarg into a union with members ap (struct argnod) and dp (struct
dolnod). It allows us to get rid of quite a few typecasts and
slightly de-obfuscate the code, as we can now easily document the
function of comarg.ap and comarg.dp in the comments in argnod.h.

Also, in two places, the initialisation to zero of individual
comarg struct members is changed to a memset(3) call. This is more
future-proof as any new members will automatically be initialised.

This commit should cause no change in behaviour and no change in the
binary representation of struct argnod.