mamake: execute(): switch from system(3) to fork(2)/execl(3)
When compiling ksh on Android/Termux (another commit is coming up
to make that possible), mamake mysteriously stopped working once it
was re-linked against libast in libast/Mamfile (see 4edfa45e).

The culprit is the system(3) library function, which simply fails
quietly on that system when mamake is linked against libast -- not
even errno is set, though there is a return status of 1.

I don't know why that is, but never mind: system(3) is deprecated
anyway, and the rest of this code base now very much requires
fork(2), so we might as well use it. (re: 7b0e0776, e7ed6790,
6605c68f, 2d674bb1, 976adfc0, ef9a5c76)

This commit switches the execute() function, which executes shell
actions, over to the standard Unix fork/exec routine.

It's even slightly more efficient this way, as system(3) invokes
everything via /bin/sh, so now we are no longer invoking our own
$SHELL via another shell.

The execl(3) call sets argv[0] to "sh" regardless of the actual
path of the shell. This put some shells, including ksh 93u+m, bash,
yash, and zsh, in POSIX compatibility mode. It should make it
possible to use zsh as the $SHELL.