Fix --showme output of multiple redirections
@adavies42 reports:
> redirects don't print properly under set -o showme:

$ ksh --showme -c ';ls >/dev/null 2>/dev/null'
+ ls
+ 1> /dev/null $ ▂

> afaict the output should be identical to the same script with
> set -x instead:

$ ksh -x -c ';ls >/dev/null 2>/dev/null'
+ ls
+ 1> /dev/null 2> /dev/null
$ ▂

Analyis: xtrace and showme are largely executed by the same code.
However, sh_redirect() returns after the first redirection if
showme is on (io.c, lines 1404-1405). It should process all the
redirections in the enclosing for loop (1168-1565), even if only to
print them without executing.

So the fix is simple: change that return to a continue.

This bug has been present since showme was introduced in ksh 93s-
2006-09-12.

Resolves: https://github.com/ksh93/ksh/issues/753