mamake: Revert broken probe() call change (re: e8b45cc8)
I messed up and broke mamake. I wrote:

>   - make(): setv: Get rid of redundant state.probed variable. Since
>     'setv' only parses the value if the variable isn't already set,
>     it's sufficient to simply call probe() upon parsing the value.

But that is not true; CC may be exported into the environment prior
to building, in which case mamake imports it as a variable before
any 'setv' gets executed, the 'setv CC' gets ignored, and (after
the referenced commit) probe() is never called, breaking the build.
So probe() needs to be called, but only once, regardless of whether
CC was already set or not; this is why we need state.probed.

Thanks to Cy Schubert (@cschuber) for reporting the breakage.

Resolves: https://github.com/ksh93/ksh/issues/754
See also: https://github.com/ksh-community/ksh/issues/20
See also: https://github.com/ksh-community/ksh/issues/21