getopts, read: use sh_scoped() to get local scope
The getopts and read built-ins use a convoluted method to get the
local scope (if any) for the OPTARG and REPLY variables,
respectively -- involving calls to nv_name() and nv_open(). But ksh
has the sh_scoped() function for this purpose, which is the standard
way of searching locally scoped variables by name and is a lot more
straightforward. This commit changes these to use that.

There should be no change in behaviour, other than perhaps a minor
performance improvement.