Fix regression involving SIGPIPE in pipe in subshell (re: d71b313a)
A rather nasty regression: when the second-to-rightmost element of
a pipeline within a subshell is interrupted by SIGPIPE (becuase the
rightmost element didn't read data written to the pipe), then the
SIGPIPE is incorrectly propagated to the main shell, which then
handles it incorrectly and emits the signal with the number of the
exit status of the rightmost command.

This problem is intermittently reproducible, depending on the
system and other circumstances. Some interactive-shell reproducers
that act incorrectly on my system:

  $ (echo OK | read bad@varname); echo $?
  -ksh: read: bad@varname: invalid variable name
  Hangup
  257

  $ (echo foo | false); echo $?
  Hangup
  257

  $ (fn(){ return 0; };echo ok|fn); echo $?
  128

  $ (fn(){ return 1; };echo ok|fn); echo $?
  Hangup
  257

  $ (fn(){ return 2; };echo ok|fn); echo $?
          # SIGINT: no output, 'echo $?' never executed

  $ (fn(){ return 3; };echo ok|fn); echo $?
  Quit
  259

  $ (fn(){ return 4; };echo ok|fn); echo $?
  Illegal instruction
  260

This problem can sometimes cause a script to abort, too. I noticed
it when the namespace.sh regression test script added in the
previous commit (c97c8995) got aborted completely upon failure of
the added tests.

This commit applies, as an urgent fix, the minimum change necessary
to eliminate the problem. Signal handling may need a deeper rethink
(again), but that will need time and thorough testing.

This regression has been in ksh 93u+m since shortly before the
1.0.0 release, so about two years. No one reported it, so I hope
not too many have been affected.

src/cmd/ksh93/sh/fault.c: sh_done():
- Exempt SIGPIPE in a real subshell from the routine that
  propagates the signal to the parent process. This simply should
  not happen in that particular case.
- Since this makes it fall back to a regular exit() whose status
  code is trimmed to 8 bits (so we cannot set the 9th bit, which
  was the point of reemitting the signal), at least make sure we
  exit with a code with the 8th bit set, to remain POSIX compliant
  with an exit status > 128 for a signal. This is done by checking
  if the local sig variable is nonzero (in which case it will,
  currently, be SIGPIPE's signal number).

src/cmd/ksh93/tests/signal.sh:
- Add regression test for this problem. It may not fail reliably on
  all systems before the fix.

src/cmd/ksh93/tests/namespace.sh:
- Make the tests added in c97c8995 abort-proof for older versions
  by launching a child shell.