Fix crash on unmatched '((' in compound command
Minimal reproducer:

    {
            (( 1
            $(( 1 ))
    }

Running that script segfaults ksh. Expected: syntax error.

Analysis: When encountering a '((', lex.c recursively invokes the
lexer to detect a closing '))' or two separate closing ')'s. This
recursive call determines whether the '((' is interpreted as
EXPRSYM, i.e. the start of an arithmetic command or expression.

The bug is that the return value of the recursive call was passed
down without a check whether the matching closing parentheses were
found. If a character classed as S_POP in the ST_NESTED state table
(see lexstates.c) was found, then the recursive call returns 0
(lex.c, line 1099). If a syntax erorr is not thrown then, the
parser ends up crashing due to a null argument pointer.

src/cmd/ksh93/sh/lex.c:
- Check whether the recusive lexer call returns the expected LPAREN
  upon encountering a matching ')', or EXPRSYM upon encountering a
  matching '))'. Throw a syntax error in every other case.

Thanks to Roland Mainz (@gisburn) for the report.
Resolves: https://github.com/ksh93/ksh/issues/764