Re-re-fix SIGPIPE subshell bug (re: 22cf8624, 3b33b438, 7e6bbf85)
Third time's a charm?

The second fix (first referenced commit) broke:

  $ (sh -c 'kill -s PIPE $$'); echo $?
  13

Expected: 269 (assuming SIGPIPE==13)
This killed the init of the modernish shell library.

src/cmd/ksh93/sh/jobs.c:
- Revert change. This is broken, too.

So now that regression from 2022-07-22 is back. My new debugging
strategy is to place an abort() call in job_wait() and use the
resulting stack trace to find out just where the job_wait() call is
located that incorrectly sets sh.chldexitstatus in the failing
signal.sh regression test (which involves a command substitution).

It turns out to be here, in subshell.c, in code that only gets
executed for command substitutions:

697:		int e = sh.exitval;
698:		job_wait(sh.spid);
699:		sh.exitval = e;

The saving and restoring of sh.exitval was added in the third
referenced commit. The problem is that job_wait may change not only
the exit status, but also sh.chldexitsig. So if sh.exitval is
restored but not sh.chldexitsig, then (if the latter is set) we
take the wrong exit status as a signal when resending it to the
parent in sh_done(). That is the bug. It wasn't job_wait()'s fault.

src/cmd/ksh93/sh/subshell.c: sh_subshell():
- Save and restore sh.chldexitsig along with sh.exitval when
  calling job_wait().

But that's not enough. I found a three more cases where the exit
status got restored from a saved value after a job_wait() call. In
all such cases, we need to make sure that sh.chldexitsig gets
restored along with sh.exitval.

src/cmd/ksh93/sh/xec.c: sh_exec():
- Save and restore sh.chldexitsig along with sh.exitval for two
  more job_wait() calls. This fixes the non-comsub intermittent
  regressions (see second referenced commit).

src/cmd/ksh93/sh/path.c: command_xargs():
- The 'command -x' code also sets sh.exitval after job_wait(),
  potentially causing it to be inconsistent with sh.chldexitsig.
  I'm taking a different approach here: simply break the loop,
  aborting the 'command -x' run, if sh.chldexitsig is set after
  job_wait(). That means one of the command subinvocations got
  interrupted by a signal, so the whole run should stop anyway.