libast: Cdt: remove dtnew() (re: 40f46650)
The dtnew() function was a Vmalloc-specific wrapper for dtopen()
which allocated a tree in a specific Vmalloc region. With Vmalloc
removed, it became entirely redundant, so it's now removed as well.
It was never documented in cdt.3.

If any code using it is backported later, the dtnew() calls should
simply be changed to dtopen() as part of changing all the Vmalloc
calls to standard malloc/calloc/etc. calls, and #include <dt.h>
should be changed to #include <cdt.h>.