libast: remove obsolete fallbacks for mem* funtions
These functions have all been either in POSIX since 1988 or in C
since C90, so, again, the fallbacks are dead code.

memccpy has an extensive feature test in features/lib to check it's
working properly, but it was broken when we inherited it (see
5d8ae360) and there is evidence that the AST folks did not have a
good understanding of undefined behaviour associated with its use
on overlapping buffers (see 8b07d2a0). So I don't quite trust that
test. Between that and its age, I think we can do without.