build: mamake: reap(): reset pid before error_making (re: 2c98350b)
The referenced commit introduced a minor bug in mamake when
throwing an error in a shell action run in parallel with others.

reap() should reset a rule's pid variable before calling
check_shellaciton(), not after.

Otherwise, check_shellaction() may call error_making() which calls
exit_wait() which (via walk()) calls reap() again, which (if r->pid
is not reset) calls waitpid() again, which will then return -1 for
error.

Which leads us to another bug in reap(): the code was accounting
for the 0 return value that waitpid() returns when passed the
WNOHANG flag and the process isn't finished yet, but not for the -1
return code when waitpid() fails. As a result, mamake was setting
an indeterminate exit status.

Checking for a return value < 1 from waitpid covers both the
WNOHANG "not yet finished" case and the error case (not that the
error case should ever occur after the pid variable fix above).