Fix error in include/wait.h (re: abc0f1a0)
I keep forgetting that in an iffe script line like

   tst ok_wif sys/types.h sys/wait.h note{ POSIX wait(2) macros ok }end compile{

there is is not just the ok_wif test, but also implicit tests for
sys/types.h and sys/wait.h, which emit _sys_types and _sys_wait
symbols if those headers exist.

The _sys_types one already isn't tested for anywhere (it's
extremely portable) but the _sys_wait one was still being tested
for in src/lib/libast/include/wait.h. This means that, after its
removal in the referenced commit, it was no longer including
<sys/wait.h>, causing build failure on FreeBSD and (I'm sure) other
systems.

Since sys/wait.h is also a extremely portable in 2024, instead of
retoring the test, this simply removes the check for it.