Simplify malloc(3)/realloc(3) usage
At a number of places in this code base, a memory block gets
allocated with malloc(3) first and then grown with realloc(3) if
necessary. But, if the pointer to the block is initialised to NULL,
then realloc(3) can always be used, because realloc(3) with a
pointer argument of NULL is equivalent to malloc(3). This allows
for a simplification at various points in the code.

We can also eliminate redundant typecasts while we're at it (these
are not needed as of the change to C90).