change all #define<TAB> to #define<SPACE>
The string '#define' is 7 positions long. For the standard tab
length of 8, that makes #define<TAB>something visually identical
to #define<SPACE>something. This gets annoying when using a
nonstandard tab length, as the two variants were mixed randomly.
This commit fixes it by changing all those tabs to spaces.