mamake: eliminate strcpy(3) and sprintf(3) use
These functions throw deprecation or security warnings on some
compilers or linkers. None of the usages are actually problematic
because the code always makes sure the actual or maximum string
length is known and fits into the buffers. However...

In mamake, strcpy is only used to copy into a buffer that was just
allocated, so the string length was already calculated with strlen.
While strcpy has to scan for the terminating zero byte, memcpy can
copy with a known length, which is faster. So this commit replaces
all strcpy use with memcpy, applying code tweaks where expedient.

sprintf may become problematic so easily that it's best to use
snprintf as an extra guard against future potential buffer
overflows. The cost is negligible. So this commit does that, too.