fun/cd: only print PWD in special cases (re: a566596f)
When the cd autoloadable function is used (to make cd play nicely
with dirs/pushd/popd), there is an unexpected user-facing change:
the new directory is printed after using cd with a normal directory
path. This might be practical for some, but it's nonstandard; cd is
not supposed to print anything to standard output in that case.
Also, most people now display the PWD in their prompts, so it's
redundant. Finally, when this cd is combined with autocd, the PWD
is already included in its 'cd -- DIR' trace output, so printing it
again is extra annoying.

src/cmd/ksh93/fun/cd:
- At the end, only print $PWD if the argument given was '-' or a
  relative or absolute directory stack number. This is the case if
  type < 3.