[v1.1] Fix case modification for unset variables (re: c1762e03)
The behaviour is incorrect if the variable is unset. For instance:
	unset x
	echo ${x^oops}
outputs 'oops' but should output an empty line.

Analysis: The varsub function in macro.c contains, on line 1689,
some advanced logic (including a bitwise xor operator '^' used on
the result of logical operations! well, it does work) to decide if
the operand/pattern string after the operator in a complex
expansion should be skipped. In the referenced commit, "c==',' ||
c=='^'" was backported here from ksh 93v-. This causes that string
never to be skipped, no matter if the variable is set or unset.

src/cmd/ksh93/sh/macro.c: varsub():
- Remove the erroneous "c==',' || c=='^'".
- Instead, if the variable is unset (v==NULL), skip the pattern
  string if newops is set to 1, and not only if the operator
  character is '#' or '%'. Note that newops is set if the operator
  starts with one of the S_MOD2 characters '/' '#' '%' '^' ',' (but
  this is irrelevant for '/' as that is checked previously).

src/cmd/ksh93/sh.1:
- While we're here, add missing documentation for the default
  pattern "?" that is used if the pattern string is omitted.